backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
coverage==7.13.5
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
execnet==2.1.2
idna==3.11
iniconfig==2.3.0
Markdown==3.10.2
oauthlib==3.3.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
pytest-xdist==3.8.0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
